backend/requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
Django==4.2.23
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
ndg-httpsclient==0.5.1
psycopg2-binary==2.9.10
pusher==3.3.3
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
pyOpenSSL==25.1.0
python-dotenv==1.1.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
